image to float;

+++ study_pytorch/HumanProteinDatasets.hpp
@@ -84,10 +84,10 @@ torch::data::Example<torch::Tensor, torch::Tensor> HumanProteinAtlasDataset::get
     cv::Mat img(512, 512, CV_8UC1);
     cv::resize(img_, img, img.size(), 0, 0, cv::INTER_AREA);
     auto image = torch::tensor(torch::ArrayRef < uint8_t > (img.data, img.rows * img.cols * 1)).view(
-            {1, img.rows, img.cols});
+            {1, img.rows, img.cols}).to(at::kFloat);
 
     //read label
-    torch::Tensor label = torch::zeros({28});
+    torch::Tensor label = torch::zeros({28}).to(at::kFloat);
     std::stringstream ss;
     ss << item.getTarget();
     while (true) {
